db: additional tweak to the sstable boundary generation
Rather than zeroing the seqnum of a range tombstone if it is the
smallest key in an sstable and it overlaps with the largest key in the
previous table, we now set its seqnum to
`prev.LargestKey.SeqNum-1`. This is mildly cleaner, though it doesn't
fix an iteration bug because `mergingIter` allows range tombstones to
shadow keys beyond the bounds of the sstables they reside in. We don't
want to change `mergingIter`'s behavior there any time soon, because
RocksDB expects that behavior.

Found via the metamorphic test.

See #479
Fixes #493